fix(signature-collection): Refactor for parliamentary applications
