--- libs/application/templates/signature-collection/parliamentary-list-creation/assets/Logo.tsx
@@ -1,47 +1,0 @@
-import React, { FC } from 'react'
-
-export const Logo: FC<React.PropsWithChildren<unknown>> = () => (
-  <svg
-    width="80"
-    height="80"
-    viewBox="0 0 80 80"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <g clipPath="url(#clip0_8686_17776)">
-      <g clipPath="url(#clip1_8686_17776)">
-        <path
-          d="M37.3702 0.566454L18.3626 9.81752C18.2539 9.87023 18.1623 9.95247 18.0982 10.0548C18.034 10.1572 18 10.2755 18 10.3962V35.9469C18 36.1175 18.0678 36.2812 18.1884 36.4018C18.3091 36.5225 18.4728 36.5903 18.6434 36.5903H37.6511C37.8217 36.5903 37.9854 36.5225 38.106 36.4018C38.2267 36.2812 38.2945 36.1175 38.2945 35.9469V1.14688C38.2943 1.0379 38.2665 0.93074 38.2137 0.835409C38.1608 0.740079 38.0847 0.659682 37.9924 0.601723C37.9002 0.543764 37.7947 0.510131 37.6859 0.503964C37.5771 0.497796 37.4685 0.519295 37.3702 0.566454Z"
-          fill="#020350"
-        />
-        <path
-          d="M37.3853 40.088L18.393 48.4965C18.2777 48.5478 18.1797 48.6313 18.1108 48.737C18.042 48.8427 18.0052 48.9661 18.0049 49.0923V79.7663C18.0048 79.8747 18.0317 79.9815 18.0833 80.0768C18.1348 80.1722 18.2094 80.2532 18.3001 80.3125C18.3909 80.3719 18.495 80.4076 18.603 80.4166C18.7111 80.4255 18.8197 80.4073 18.9189 80.3638L37.9113 72.0233C38.0273 71.9724 38.1259 71.8888 38.1951 71.7827C38.2644 71.6765 38.3012 71.5526 38.3011 71.4259V40.6906C38.3021 40.5815 38.2758 40.4739 38.2245 40.3776C38.1732 40.2814 38.0986 40.1995 38.0075 40.1395C37.9164 40.0796 37.8117 40.0435 37.703 40.0345C37.5943 40.0254 37.485 40.0439 37.3853 40.088Z"
-          fill="#020350"
-        />
-        <path
-          d="M41.6016 60.1664C41.4278 60.1664 41.2611 60.0973 41.1382 59.9744C41.0153 59.8515 40.9463 59.6849 40.9463 59.5111V37.2455C40.9463 37.1595 40.9632 37.0743 40.9962 36.9947C41.0291 36.9152 41.0774 36.843 41.1382 36.7821C41.1991 36.7213 41.2713 36.673 41.3508 36.6401C41.4303 36.6072 41.5156 36.5902 41.6016 36.5902H60.5871C60.6732 36.5902 60.7584 36.6072 60.8379 36.6401C60.9174 36.673 60.9897 36.7213 61.0505 36.7821C61.1114 36.843 61.1596 36.9152 61.1926 36.9947C61.2255 37.0743 61.2425 37.1595 61.2425 37.2455V40.5255C61.2425 45.7346 59.1732 50.7303 55.4898 54.4137C51.8064 58.0971 46.8107 60.1664 41.6016 60.1664Z"
-          fill="#020350"
-        />
-        <path
-          d="M51.0942 12.2496C53.2302 12.2496 55.3183 12.883 57.0944 14.0698C58.8704 15.2565 60.2547 16.9432 61.0721 18.9167C61.8895 20.8901 62.1034 23.0616 61.6867 25.1566C61.27 27.2516 60.2414 29.176 58.7309 30.6864C57.2205 32.1968 55.2962 33.2254 53.2012 33.6421C51.1062 34.0588 48.9347 33.845 46.9612 33.0275C44.9878 32.2101 43.301 30.8258 42.1143 29.0498C40.9276 27.2737 40.2942 25.1857 40.2942 23.0496C40.2942 20.1853 41.432 17.4383 43.4574 15.4129C45.4828 13.3875 48.2299 12.2496 51.0942 12.2496Z"
-          fill="#020350"
-        />
-      </g>
-    </g>
-    <defs>
-      <clipPath id="clip0_8686_17776">
-        <rect width="80" height="80" fill="white" />
-      </clipPath>
-      <clipPath id="clip1_8686_17776">
-        <rect
-          width="44.2553"
-          height="80"
-          fill="white"
-          transform="translate(18 0.5)"
-        />
-      </clipPath>
-    </defs>
-  </svg>
-)
-
-export default Logo
