--- libs/application/templates/signature-collection/parliamentary-list-creation/src/fields/index.tsx
@@ -1,1 +1,0 @@
-export { CopyLink } from './CopyLink'
