+++ libs/application/templates/signature-collection/parliamentary-list-creation/src/forms/Done.ts
@@ -4,85 +4,73 @@ import {
   buildSection,
   buildMessageWithLinkButtonField,
   buildDescriptionField,
-  buildCustomField,
+  buildImageField,
+  buildCopyLinkField,
+  getValueViaPath,
 } from '@island.is/application/core'
 import { Form, FormModes } from '@island.is/application/types'
 import { m } from '../lib/messages'
+import Jobs from '@island.is/application/templates/signature-collection/assets/Jobs'
 
 export const Done: Form = buildForm({
   id: 'done',
   mode: FormModes.COMPLETED,
   children: [
-    /* Sections for the stepper */
-    buildSection({
-      id: 'screen1',
-      title: m.dataCollection,
-      children: [],
-    }),
-    buildSection({
-      id: 'screen2',
-      title: m.information,
-      children: [],
-    }),
-    buildSection({
-      id: 'screen3',
-      title: m.overview,
-      children: [],
-    }),
-    /* ------------------------ */
     buildSection({
       id: 'doneScreen',
-      title: m.listCreated,
       children: [
         buildMultiField({
           id: 'doneScreen',
           title: m.listCreated,
           description: m.listCreatedDescription,
           children: [
-            buildDescriptionField({
-              id: 'nextStepsTitle',
-              title: m.nextSteps,
-              titleVariant: 'h3',
-              marginBottom: 1,
+            buildImageField({
+              id: 'doneImage',
+              image: Jobs,
+              imageWidth: 'auto',
+              imagePosition: 'center',
+              marginTop: 'none',
             }),
-            //Set up separately for even spacing
             buildDescriptionField({
               id: 'nextStepsDescription',
               description: m.nextStepsDescription,
               titleVariant: 'h4',
               marginBottom: 5,
             }),
-            buildCustomField({
+            buildCopyLinkField({
               id: 'copyLink',
-              component: 'CopyLink',
+              link: ({ externalData }) =>
+                getValueViaPath(externalData, 'submit.data.slug') ??
+                'https://island.is/',
+              semiBoldLink: true,
+              marginBottom: 'none',
             }),
             buildMessageWithLinkButtonField({
-              condition: (_, externalData) => {
-                return !(
-                  externalData?.delegatedToCompany.data as {
-                    delegatedToCompany: boolean
-                  }
-                ).delegatedToCompany
-              },
+              condition: (_, externalData) =>
+                getValueViaPath(
+                  externalData,
+                  'delegatedToCompany.data.delegatedToCompany',
+                ) === false,
               id: 'done.goToServicePortal',
-              title: '',
               url: '/minarsidur/min-gogn/listar/althingis-medmaelasofnun',
               buttonTitle: m.linkFieldButtonTitle,
               message: m.linkFieldMessage,
             }),
             buildMessageWithLinkButtonField({
-              condition: (_, externalData) => {
-                return (
-                  externalData?.delegatedToCompany.data as {
-                    delegatedToCompany: boolean
-                  }
-                ).delegatedToCompany
-              },
+              condition: (_, externalData) =>
+                getValueViaPath(
+                  externalData,
+                  'delegatedToCompany.data.delegatedToCompany',
+                ) === true,
               id: 'done.goToServicePortalCompany',
-              title: '',
               url: '/minarsidur/fyrirtaeki/listar/althingis-medmaelasofnun',
               buttonTitle: m.linkFieldButtonCompanyTitle,
               message: m.linkFieldCompanyMessage,
+              marginBottom: 'containerGutter',
+            }),
+            buildDescriptionField({
+              id: 'space',
+              space: 'containerGutter',
             }),
           ],
         }),
