+++ libs/application/templates/signature-collection/parliamentary-list-creation/src/forms/Prerequisites.ts
@@ -5,7 +5,7 @@ import {
   buildSubmitField,
   buildSection,
   buildMultiField,
-  buildDescriptionField,
+  buildImageField,
 } from '@island.is/application/core'
 import {
   DefaultEvents,
@@ -15,13 +15,13 @@ import {
 } from '@island.is/application/types'
 
 import { m } from '../lib/messages'
-import Logo from '../../assets/Logo'
 import {
   ParliamentaryCollectionApi,
-  CandidateApi,
   ParliamentaryIdentityApi,
   IsDelegatedToCompanyApi,
 } from '../dataProviders'
+import DigitalServices from '@island.is/application/templates/signature-collection/assets/DigitalServices'
+import Logo from '@island.is/application/templates/signature-collection/assets/Logo'
 
 export const Prerequisites: Form = buildForm({
   id: 'CreateListPrerequisites',
@@ -32,15 +32,18 @@ export const Prerequisites: Form = buildForm({
   children: [
     buildSection({
       id: 'intro',
-      title: m.intro,
       children: [
         buildMultiField({
           id: 'intro',
           title: m.introTitle,
           description: m.introDescription,
           children: [
-            buildDescriptionField({
-              id: 'introText',
+            buildImageField({
+              id: 'doneImage',
+              image: DigitalServices,
+              imageWidth: 'auto',
+              imagePosition: 'center',
+              marginTop: 'none',
             }),
           ],
         }),
@@ -48,7 +51,6 @@ export const Prerequisites: Form = buildForm({
     }),
     buildSection({
       id: 'approveExternalData',
-      title: m.dataCollection,
       children: [
         buildExternalDataProvider({
           id: 'approveExternalData',
@@ -83,11 +85,12 @@ export const Prerequisites: Form = buildForm({
               title: '',
               subTitle: '',
             }),
-            buildDataProviderItem({
+            // Todo: Add back once needed
+            /*buildDataProviderItem({
               provider: CandidateApi,
               title: '',
               subTitle: '',
-            }),
+            }),*/
             buildDataProviderItem({
               provider: IsDelegatedToCompanyApi,
               title: '',
@@ -96,26 +99,6 @@ export const Prerequisites: Form = buildForm({
           ],
         }),
       ],
-    }),
-    buildSection({
-      id: 'screen3',
-      title: m.information,
-      children: [],
-    }),
-    buildSection({
-      id: 'screen4',
-      title: m.constituency,
-      children: [],
-    }),
-    buildSection({
-      id: 'screen6',
-      title: m.overview,
-      children: [],
-    }),
-    buildSection({
-      id: 'screen7',
-      title: m.listCreated,
-      children: [],
     }),
   ],
 })
