+++ libs/application/templates/signature-collection/parliamentary-list-creation/src/index.ts
@@ -2,7 +2,6 @@ import createCollectionTemplate from './lib/createCollectionTemplate'
 import { z } from 'zod'
 import { dataSchema } from './lib/dataSchema'
 
-export const getFields = () => import('./fields')
 export * from './lib/errors'
 
 export default createCollectionTemplate
