+++ libs/application/templates/signature-collection/parliamentary-list-creation/src/lib/constants.ts
@@ -11,27 +11,6 @@ export enum Roles {
   APPLICANT = 'applicant',
 }
 
-export type SubmitResponse = {
-  success: boolean
-  orderId?: string[]
-}
-
 export enum ApiActions {
   submitApplication = 'submit',
-}
-
-export type Manager = {
-  manager: {
-    name: string
-    nationalId: string
-  }
-  constituency: string
-}
-
-export type Supervisor = {
-  supervisor: {
-    name: string
-    nationalId: string
-  }
-  constituency: string
 }
