+++ libs/application/templates/signature-collection/parliamentary-list-creation/src/lib/createCollectionTemplate.ts
@@ -134,11 +134,12 @@ const createListTemplate: ApplicationTemplate<
           status: 'completed',
           progress: 1,
           lifecycle: pruneAfterDays(30),
-          onEntry: defineTemplateApi({
+          //Todo: add back once needed
+          /*onEntry: defineTemplateApi({
             action: ApiActions.submitApplication,
             shouldPersistToExternalData: true,
             throwOnError: true,
-          }),
+          }),*/
           roles: [
             {
               id: Roles.APPLICANT,
