--- libs/application/templates/signature-collection/parliamentary-list-creation/src/lib/utils.ts
@@ -1,5 +1,0 @@
-import { formatPhoneNumber } from '@island.is/application/ui-components'
-
-export const formatPhone = (phone: string) => {
-  return formatPhoneNumber(phone.replace(/(^00354|^\+354|\D)/g, ''))
-}
